refactor(shell): sort WMI properties alphabetically
- Add alphabetical sorting to `QueryUtility.getPropertiesFromSerializedNameAnnotation()`.
- Update `QueryUtilityTest` to reflect the new property order.
- Refactor test file paths for `ScriptUtilityTest.java` and `QueryUtilityTest.java`.
